--- src/main/java/com/System/clinic/ClinicApplication.java
@@ -1,13 +1,0 @@
-package com.System.clinic;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ClinicApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ClinicApplication.class, args);
-	}
-
-}
