--- src/main/java/com/System/clinic/Dto/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.System.clinic.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorResponse {
-    private String field;
-    private String message;
-
-}
