--- src/main/java/com/System/clinic/Dto/Page.java
@@ -1,5 +1,0 @@
-package com.System.clinic.Dto;
-
-public record Page(String title) {
-
-}
