--- src/main/java/com/System/clinic/Dto/PageRequest.java
@@ -1,14 +1,0 @@
-package com.System.clinic.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PageRequest {
-    private int page;
-    private int pageSize;
-
-}
