--- src/main/java/com/System/clinic/Dto/ProfissionalDTO.java
@@ -1,40 +1,0 @@
-package com.System.clinic.Dto;
-
-import java.time.LocalDate;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ProfissionalDTO {
-
-    private Long idProfissional;
-
-    @NotBlank
-    private String name;
-
-    @NotNull
-    private LocalDate dataNascimento;
-
-    @NotBlank
-    private String especialidade;
-
-    @NotBlank
-    private String telefone;
-
-    @NotBlank
-    private String senha;
-
-    @NotBlank
-    private String cns;
-
-    @NotBlank
-    private String cpf;
-
-    @NotNull
-    private String email;
-
-}
