--- src/main/java/com/System/clinic/entity/agendamentoEntity.java
@@ -1,5 +1,0 @@
-package com.System.clinic.entity;
-
-public class agendamentoEntity {
-
-}
