--- src/main/java/com/System/clinic/exeption/NotFoundExeption.java
@@ -1,5 +1,0 @@
-package com.System.clinic.exeption;
-
-public class NotFoundExeption extends RuntimeException {
-
-}
