--- src/main/java/com/System/clinic/mapping/agendamentoMapping.java
@@ -1,5 +1,0 @@
-package com.System.clinic.mapping;
-
-public class agendamentoMapping {
-
-}
