--- src/main/java/com/System/clinic/repository/agendamentoRepository.java
@@ -1,5 +1,0 @@
-package com.System.clinic.repository;
-
-public class agendamentoRepository {
-
-}
