--- src/main/java/com/System/clinic/util/DateTimeFormatterUtils.java
@@ -1,12 +1,0 @@
-package com.System.clinic.util;
-
-import java.time.format.DateTimeFormatter;
-
-public class DateTimeFormatterUtils {
-
-    public static DateTimeFormatter FORMATO_DATA_BRASIL = DateTimeFormatter.ofPattern("dd/MM/yyyy");
-
-    private DateTimeFormatterUtils() {
-    }
-
-}
