+++ src/main/java/com/system/clinic/ClinicApplication.java
@@ -1,0 +1,13 @@
+package com.system.clinic;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ClinicApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ClinicApplication.class, args);
+	}
+
+}
