+++ src/main/java/com/system/clinic/dto/ErrorResponse.java
@@ -1,0 +1,14 @@
+package com.system.clinic.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorResponse {
+    private String field;
+    private String message;
+
+}
