+++ src/main/java/com/system/clinic/dto/Page.java
@@ -1,0 +1,5 @@
+package com.system.clinic.dto;
+
+public record Page(String title) {
+
+}
