+++ src/main/java/com/system/clinic/dto/PageRequest.java
@@ -1,0 +1,14 @@
+package com.system.clinic.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PageRequest {
+    private int page;
+    private int pageSize;
+
+}
