+++ src/main/java/com/system/clinic/entity/StatusAgendamento.java
@@ -1,0 +1,8 @@
+package com.system.clinic.entity;
+
+public enum StatusAgendamento {
+    AGENDADA,
+    CONFIRMADA,
+    CANCELADA,
+    REALIZADA
+}
