+++ src/main/java/com/system/clinic/exception/NotFoundException.java
@@ -1,0 +1,5 @@
+package com.system.clinic.exception;
+
+public class NotFoundException extends RuntimeException {
+
+}
