+++ src/main/java/com/system/clinic/mapping/AgendamentoMapping.java
@@ -1,0 +1,5 @@
+package com.system.clinic.mapping;
+
+public class AgendamentoMapping {
+
+}
