+++ src/main/java/com/system/clinic/repository/AgendamentoRepository.java
@@ -1,0 +1,5 @@
+package com.system.clinic.repository;
+
+public class AgendamentoRepository {
+
+}
