+++ src/main/java/com/system/clinic/service/ConsultaService.java
@@ -1,0 +1,28 @@
+package com.system.clinic.service;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+
+import org.springframework.stereotype.Service;
+
+import com.system.clinic.dto.AgendamentoDTO;
+
+@Service
+public interface ConsultaService {
+
+AgendamentoDTO buscarConsultaPorId(Long idConsulta);
+List<AgendamentoDTO> listarConsultas();
+List<AgendamentoDTO> ConsultasPorCns(String pacientecns);
+List<AgendamentoDTO> ConsultasPorProfissional(String profissionalNome);
+List<AgendamentoDTO> ConsultasPorData(LocalDate data);
+List<AgendamentoDTO> ConsultasPorEspecialidade(String especialidade);
+List<AgendamentoDTO> ConsultasPorHora(LocalTime horaConsulta);
+
+AgendamentoDTO save(AgendamentoDTO consultaDTO);
+void remove(Long idConsulta);
+
+
+
+
+}
