+++ src/main/java/com/system/clinic/service/ProfissionalService.java
@@ -1,0 +1,20 @@
+package com.system.clinic.service;
+
+
+import java.util.List;
+
+import org.springframework.stereotype.Service;
+
+import com.system.clinic.dto.ProfissionalDTO;
+
+@Service
+public interface ProfissionalService {
+
+    ProfissionalDTO findOne(Long idProfissional);
+    List<ProfissionalDTO> findAll();
+    List<ProfissionalDTO> ConsultarPorProfissional(String profissionalNome);
+    List<ProfissionalDTO> listarProfissionalPorEspecialidade(String especialidade);
+
+    ProfissionalDTO save(ProfissionalDTO profissionalDTO);
+    void remove(Long idProfissional);
+}
