+++ src/main/java/com/system/clinic/service/impl/AgendamentoServiceImpl.java
@@ -1,0 +1,1 @@
+package com.system.clinic.service.impl;
