+++ src/main/java/com/system/clinic/util/DateTimeFormatterUtils.java
@@ -1,0 +1,12 @@
+package com.system.clinic.util;
+
+import java.time.format.DateTimeFormatter;
+
+public class DateTimeFormatterUtils {
+
+    public static DateTimeFormatter FORMATO_DATA_BRASIL = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+
+    private DateTimeFormatterUtils() {
+    }
+
+}
